Refactor c-style uint casts to use static cast macro

+++ src/common/src/bus.cpp
@@ -3,10 +3,10 @@
 
 uint16_t BUS::read_u16(Address const *address) {
   auto bytes = this->read(address, sizeof_ui16);
-  return (uint16_t)bytes;
+  return bytes;
 }
 
 uint8_t BUS::read_u8(Address const *address) {
   auto bytes = this->read(address, sizeof_ui8);
-  return (uint8_t)bytes;
+  return bytes;
 }
